Revert "Bug fixes (#101)"
This reverts commit bedb363a806273dfb4942b450e1b8f81ff1651f6.
