+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "preact-material-components",
-  "version": "1.0.2",
+  "version": "1.0.1",
   "description": "preact wrapper for \"Material Components for the web\"",
   "module": "index.js",
   "main": "dist/index.js",
